Proyecto1_Fase2/server-python/server-python/requirements.txt
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
mysqlclient==2.1.1
redis==5.0.1